Migrated expiration bar and expiration descp. from account fee settings - deleted account status and base fee from settings - migrated learn more link
added missed comment

added changelog entry

Fix: render correct currency code in expiration bar and description (review feedback)

Fix: BTypo Base Fee was rendered as BaseFee
