+++ client/components/account-status/account-fees/expiration-bar.js
@@ -1,0 +1,28 @@
+/** @format */
+
+/**
+ * Internal dependencies
+ */
+import ProgressBar from 'components/progress-bar';
+import { formatCurrency } from 'utils/currency';
+
+const ExpirationBar = ( {
+	feeData: {
+		volume_allowance: volumeAllowance,
+		current_volume: currentVolume,
+	},
+	currencyCode,
+} ) => {
+	if ( ! volumeAllowance ) {
+		return null;
+	}
+	return (
+		<ProgressBar
+			progressLabel={ formatCurrency( currentVolume, currencyCode ) }
+			totalLabel={ formatCurrency( volumeAllowance, currencyCode ) }
+			progress={ currentVolume / volumeAllowance }
+		/>
+	);
+};
+
+export default ExpirationBar;
