+++ client/components/account-status/account-fees/expiration-description.js
@@ -1,4 +1,5 @@
 /** @format */
+
 /**
  * External dependencies
  */
@@ -9,33 +10,11 @@ import moment from 'moment';
 /**
  * Internal dependencies
  */
-import ProgressBar from 'components/progress-bar';
-import {
-	formatAccountFeesDescription,
-	getCurrentFee,
-} from 'utils/account-fees';
 import { formatCurrency } from 'utils/currency';
-
-const ExpirationBar = ( {
-	feeData: {
-		volume_allowance: volumeAllowance,
-		current_volume: currentVolume,
-	},
-} ) => {
-	if ( ! volumeAllowance ) {
-		return null;
-	}
-	return (
-		<ProgressBar
-			progressLabel={ formatCurrency( currentVolume ) }
-			totalLabel={ formatCurrency( volumeAllowance ) }
-			progress={ currentVolume / volumeAllowance }
-		/>
-	);
-};
 
 const ExpirationDescription = ( {
 	feeData: { volume_allowance: volumeAllowance, end_time: endTime },
+	currencyCode,
 } ) => {
 	let description;
 	if ( volumeAllowance && endTime ) {
@@ -45,7 +24,7 @@ const ExpirationDescription = ( {
 				'Discounted base fee expires after the first %1$s of total payment volume or on %2$s.',
 				'woocommerce-payments'
 			),
-			formatCurrency( volumeAllowance ),
+			formatCurrency( volumeAllowance, currencyCode ),
 			dateI18n( 'F j, Y', moment( endTime ).toISOString() )
 		);
 	} else if ( volumeAllowance ) {
@@ -55,7 +34,7 @@ const ExpirationDescription = ( {
 				'Discounted base fee expires after the first %1$s of total payment volume.',
 				'woocommerce-payments'
 			),
-			formatCurrency( volumeAllowance )
+			formatCurrency( volumeAllowance, currencyCode )
 		);
 	} else if ( endTime ) {
 		description = sprintf(
@@ -72,30 +51,4 @@ const ExpirationDescription = ( {
 	return <p className="description">{ description }</p>;
 };
 
-const AccountFees = ( { accountFees } ) => {
-	const currentFee = getCurrentFee( accountFees );
-	const feeDescription = formatAccountFeesDescription( accountFees );
-
-	return (
-		<>
-			<p>{ feeDescription }</p>
-			<ExpirationBar feeData={ currentFee } />
-			<ExpirationDescription feeData={ currentFee } />
-			<p>
-				<a
-					href={
-						accountFees.discount.length
-							? 'https://woocommerce.com/terms-conditions/woocommerce-payments-promotion/'
-							: 'https://docs.woocommerce.com/document/payments/faq/fees/'
-					}
-					target="_blank"
-					rel="noopener noreferrer"
-				>
-					{ __( 'Learn more', 'woocommerce-payments' ) }
-				</a>
-			</p>
-		</>
-	);
-};
-
-export default AccountFees;
+export default ExpirationDescription;
